refactor: extract review photos to a separate component
